@@ -1,17 +1,17 @@
 /* Copyright 2013 Ivan Iljkic
-*
-* Licensed under the Apache License, Version 2.0 (the "License"); you may not
-* use this file except in compliance with the License. You may obtain a copy of
-* the License at
-*
-* http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-* WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-* License for the specific language governing permissions and limitations under
-* the License.
-*/
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
 package net.utp4j.channels.futures;
 
 import java.nio.ByteBuffer;
@@ -19,40 +19,41 @@ import java.util.concurrent.locks.ReentrantLock;
 
 /**
  * Read future.
- * @author Ivan Iljkic (i.iljkic@gmail.com)
  *
+ * @author Ivan Iljkic (i.iljkic@gmail.com)
  */
 public class UtpReadFuture extends UtpBlockableFuture {
-	
-	public UtpReadFuture() throws InterruptedException {
-		super();
-	}
 
-	protected volatile ByteBuffer buffer;
-	protected volatile UtpReadListener listener;
-	protected final ReentrantLock listenerLock = new ReentrantLock();
-	
-	public ByteBuffer getBuffer() {
-		return buffer;
-	}
-	
-	/**
-	 * Sets a listener that will be informed once the future task is completed. 
-	 */
-	public void setListener(UtpReadListener listener) {
-		listenerLock.lock();
-		this.listener = listener;
-		listenerLock.unlock();
-		
-	}
-	/**
-	 * @return The number of bytes that was read by this task.
-	 */
-	public int getBytesRead() {
-		if (buffer != null) {
-			return buffer.position();
-		}
-		return 0;
-	}
-	
+    public UtpReadFuture() throws InterruptedException {
+        super();
+    }
+
+    protected volatile ByteBuffer buffer;
+    protected volatile UtpReadListener listener;
+    protected final ReentrantLock listenerLock = new ReentrantLock();
+
+    public ByteBuffer getBuffer() {
+        return buffer;
+    }
+
+    /**
+     * Sets a listener that will be informed once the future task is completed.
+     */
+    public void setListener(UtpReadListener listener) {
+        listenerLock.lock();
+        this.listener = listener;
+        listenerLock.unlock();
+
+    }
+
+    /**
+     * @return The number of bytes that was read by this task.
+     */
+    public int getBytesRead() {
+        if (buffer != null) {
+            return buffer.position();
+        }
+        return 0;
+    }
+
 }
