@@ -1,33 +1,33 @@
 /* Copyright 2013 Ivan Iljkic
-*
-* Licensed under the Apache License, Version 2.0 (the "License"); you may not
-* use this file except in compliance with the License. You may obtain a copy of
-* the License at
-*
-* http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-* WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-* License for the specific language governing permissions and limitations under
-* the License.
-*/
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
 package net.utp4j.channels.impl.recieve;
 
 import java.net.DatagramPacket;
 
 public class UtpPassPacket extends Thread implements Runnable {
 
-	private UtpPacketRecievable reciever;
-	private DatagramPacket packet;
-	
-	public UtpPassPacket(DatagramPacket packet, UtpPacketRecievable reciever) {
-		this.packet = packet;
-		this.reciever = reciever;
-	}
-	
-	@Override
-	public void run() {
-		reciever.recievePacket(packet);
-	}
+    private final UtpPacketRecievable receiver;
+    private final DatagramPacket packet;
+
+    public UtpPassPacket(DatagramPacket packet, UtpPacketRecievable receiver) {
+        this.packet = packet;
+        this.receiver = receiver;
+    }
+
+    @Override
+    public void run() {
+        receiver.recievePacket(packet);
+    }
 }
